[Feature #159102801 Restructed Api and isolated Models and Resources
