+++ Api/.gitignore
@@ -1,0 +1,2 @@
+.pytest_cache
+/__pycache__
